requirements.txt
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2
Pillow==10.2.0

facenet-pytorch==2.6.0
resemblyzer==0.1.1.dev0
vosk==0.3.45
sounddevice==0.4.6